Nyil javitva player, nyil kesz

+++ SDLProject/include/Shot_Arrow.hpp
@@ -11,8 +11,8 @@ private:
 	Vector2D Arrow_LastSafePosition;
 	SpriteAnimation* Arrow_SpriteAnimation;
 
-	int OriginalX = -1;
-	int OriginalY = -1;
+	double OriginalX = -1;
+	double OriginalY = -1;
 
 	int TargetPosX{};
 	int TargetPosY{};
